[unit-tests] refactor fixtures into seperate folder
